fix(Chat): ip to long

+++ src/server/Chat.Infrastructure/MessageHandlers/Data/GetMessageHubItem.cs
@@ -4,7 +4,7 @@
 public class GetMessageHubItem
 {
     // ReSharper disable once InconsistentNaming
-    public int IPv4 { get; set; } = default;
+    public long IPv4 { get; set; } = default;
     public int Port { get; set; } = default;
     public string Text { get; set; } = "";
 }
